Refactored upsert openstack security group operation.
Moved business logic out of client provider and into the operation. The
client provider now just is a proxy for individual calls to the
Openstack API.

OpenstackResourceNotFoundException is now provider exception.

Moved around the resource not found exception to be more useful for the
client provider and catching it in the operation.