fix(provider/kubernetes): Cross Account Clone
Fixes cloning Server Groups across accounts. The previous implementation
was relying on the credentials of the target account which meant
cloning server groups that didn't exist in the target account failed.
This adds a `sourceCredentials` field to the Clone description that is
used to verify the source server group against the correct account.

Fixes spinnaker/spinnaker#1829