refactor(dynomite): Consolidating cache & hash structs into hashsets
Now that we're running Dynomite in staging and performing loadtests,
we've been seeing dramatically reduced performance due, in part, to
the aggregate overhead of performing individual network calls for
each key, whereas in vanilla Redis, we were able to multi-key op and
pipeline requests.

This is a swing back in the direction of larger keys, in an effort
to reduce the overhead of network call volume.