feat(core): Version provider accounts/operations
This is intended for the v2 k8s provider.

The idea is that accounts can expose a "version" which converters,
validators, and (in the future) caching agents will use to determine
when to be run. This is also useful if we want to start moving features
into release tracks, as we can flag operations with a "beta" version
that requires accounts to be marked with that version.

Sorry for the half-finished k8s work, I wanted to submit this before the
PR got too big