fix(provider/google): Handling edge cases with user data files better.
This incorporates 3 changes/improvements for handling user data files. It
allows blank lines in the user data file. It also interupts the lack of an
equal sign as a key with no value. Finally, it allows values to contain equal
signs, in other words, it only splits on the first equal sign.