fix(aws) : task stuck on wait for up instances
We are having trouble with tasks "waiting for up instances" in aws.
I think it is because this part of the code is getting throttled,
but im not possitive of the fix. This code it meant to catch the
throttling error, and stop the current task that is ruuning,
so the user can restart the task if possible.

-e