feat(provider/aws): Support expected capacity constraint when resizing
Support a constraint wherein the server groups' current capacity must
match expected prior to a resize.

There are scenarios wherein a user initiates a resize via the UI at the
same time that an autoscaling event is happening behind the scenes. In
such a scenario, the resize via UI != autoscaling and some churn will
occur.

`deck` will be changed to supply the capacity constraint.