fix(provider/kubernetes): registry init fix
for real this time. the previous fix caused registry namespaces to get
set to an empty list the second time the initialization occurred. moving
it down into this method allows us to still use the full list of
namespaces (`allNamespaces`) and setup the affected ones to be
initialized. This would occur if a new namespace was added, for
instance.