fix(provider/aws): Avoid being marked unhealthy after initialization
Avoid being marked unhealthy should there be a temporary blip
in connectivity to one or more accounts.

This frequently occurs as a result of an AWS rate limit being hit.

The `health.amazon.errors` metric can be used to alert on possible
errors that would have previously caused an instance to go unhealthy.