fix(provider/kubernetes): v2 Set default namespace dynamically
Set the default namespace according to the `kubectl` rules defined in
https://github.com/kubernetes/kubernetes/blob/bd4d511a40e142ee65edff3b286e57de502aa790/pkg/kubectl/cmd/util/factory_client_access.go#L127

In short, the default namespace is set in this way:

- If `/var/run/secrets/kubernetes.io/serviceaccount/namespace` exists,
  assume we are running in a Kubernetes container, and set the default
  namespace to the contents of this file.
- Set the default namespace to the namespace of current `kubectl`
  config context, if it exists. `kubectl config view` is used to get
  this namespace, which takes care of the complicated merges and
  overrides.
- Otherwise, set the default namespace to "default".

This fixes the issue of manifests without a namespace specification
being deployed in the namespace of clouddriver when Spinnaker has been
deployed in Kubernetes.

Previously manifests without a namespace were assumed to belong to the
"default" namespace, but clouddriver would deploy them in the
"spinnaker" namespace, where it is deployed by default by Halyard.
Deploying such manifests would cause Spinnaker to get stuck waiting
for a deployment to show up in the "default" namespace, since
clouddriver would instead have created a deployment in the "spinnaker"
namespace.