fix(provider/kubernetes): v2 Make workloads authoritative for apps
Let workload deployments have full control over (generated)
applications coming and going.

This is in line with https://github.com/spinnaker/clouddriver/pull/2303
and fixes https://github.com/spinnaker/spinnaker/issues/2259

In particular, applications will not be shown in the deck
"Applications" view if the replicaset/daemonset/deployment/statefulset
of the application has the `caching.spinnaker.io/ignore="true"`
annotation, and will be removed from the list if the annotation added
to an existing workload.