fix(provider/kubernetes): v1 eventual consistency bug in disable
It turns out it was possible that we would

1. apply an annotation
2. toggle replica set labels
3. fail to wait for the label change to be applied

Part 3. could happen because the generation of the replica set returned
from the "toggle label" operation was that of the "apply annotation"
operation. The "wait for consistency" step we usually do here was then
skipped causing us to edit pod labels before the replica set owning them
was finised updating.

This doesn't happen often, but enough to be a real problem.