fix(aws): Continue if launch configuration already exists
It's possible that a launch configuration can be created but subsequently
fail due to a network blip.

When this happens, AWS will reject any subsequent requests with an
`AlreadyExistsException` exception.

In such situations, it makes sense to continue (launch configurations
are timestamped down to the second) rather than fail the task.