perf(caching): Support retrieving an individual on-demand record by id
Currently `orca` fetches all on-demand records and filters client-side
down to the specific record it is interested in.

This PR does two things:
- returns a set of typed identifiers as part of the original force cache refresh api
- allows for an `id` parameter to be supplied to the `/cache/{cloudProvider}/{type}` api

It currently supports a single `id` parameter as the expected use case
will be as part of the `ServerGroupForceCacheRefreshTask` which deals
with one server group at a time.