fix(artifacts): s3 improvements
this commit fixes two issues with the s3 artifact provider. 1, it adds
support for configuring an s3 region. this is inline with how we already
handle it in front50, allowing for usage of buckets in separate regions.
also, when using `apiEndpoint` we not set `pathStyleAccessEnabled` which
enables support for things like Minio to act as a provider.