fix(projects): remove parallel cluster lookups
A project with a large number of applications can significantly load
the redis cache by requesting all application clusters in parallel.

This removes the Rx code that was parallelizing the lookups in favor
of a linear lookup through all the project's applications, and
additionally uses the RequestQueue when performing the cluster lookups
in case we are under load and the requests are timing out.