perf(core): Attempted move away from fetching _every_ possible Project cluster
The current approach to fetching clusters for a Project is quite
inefficient, particular for large applications wherein only a subset
of clusters are included in the project config.

This PR initially fetches cluster summaries and then applies a filter
to determine applicability. The applicable clusters are then fetched in
detail.