feat(core): General purpose `DescriptionAuthorizer`
Ported the `authorize()` method from `DescriptionValidator` and added
a metric in the event that a description presents an account _but_ no
application(s).

This metric will emit some false positives but still prove useful in the
short-term to highlight descriptions that might not implement
`ApplicationNameable` (or not implement it properly).

The previous behavior expected a `Description` to implement
`DescriptionValidator` in order to have authorization applied, however
this is not a requirement and missed some legitimate scenarios.