refactor(provider/cf): added test for deploy converter.
The intention here was to describe existing behavior. Only minor changes were made to improve testability  but not impact behavior.
CloudFoundryClient was turned into an interface, and a mocked impl was introduced that could be shared outside the ops package.
Some logic in the converter was refactored into its own method for testability.