feat(provider/kubernetes): 'liveManifestCall' mode
Been meaning to submit this for us to try out:

This is in service of faster pipeline executions, which ignore force
cache refreshes, and only serve manifest status from live calls.

Need to find a cleaner top-level way to distinguish between object
providers that query the cache vs. live data.

Important to note: this leaves caching (& the clusters tab) intact for
now, those can be disabled separately in a future PR.