feature(aws/cloudformation): Add cloudformation caching agent and provider
This patch adds two components necessary to specify and implement a
cloud formation stage in a pipeline.

* Cloudformation caching agent: it's responsible to periodically query
aws to cache the relevant information from a cloudformation stack,
including the status whenever it's in an errored state.

* Cloudformation provider: it provides an API to query the cache and get
a cloudformation stack given a stackId or list existing stacks for an
aws account and (optionally) a region.

The cloudformation controller that will be used by Orca as well as the
task that deploys this cloud formation will be added in separate PRs.

Signed-off-by: Xavi León <xavi.leon@schibsted.com>