feat(aws/cloudformation): Allow on demand cache request/invalidation
This patch enabled the CloudFormationCachingAgent to also handle on
demand request, so items of the cache are force refreshed. We need
to invalidate this cache/forcefully refresh it when we update an
existing cloudformation stack, so we track the progress of the new
operation, not the result of the previous one.

Signed-off-by: Xavi León <xavi.leon@schibsted.com>