feat(core): `DeploymentResult` support for returning target capacity
AWS deployments are somewhat special in that they create an initial
asg of size 0/0/0 and then subsequently resize it to the target
capacity.

This causes problems if the `ForceCacheRefresh` task as part of a deploy
were to ever be removed from `orca` as it's possible for the 0/0/0 asg
to be cached and the `WaitForUpInstances` task to prematurely complete.

I'd like to remove the `ForceCacheRefresh` task as part of a deploy,
hence this PR and a follow-up in `orca` that makes it aware of the
capacity being passed back from `clouddriver`.