fix(appengine): Fix appengine deployments
The new JobExecutorLocal is in Java, and fails when passed an
array containing a mixture of Strings and GStrings.  As a GString
is not a subclass of String (which is final) it is incorrect to
create a collection containing a mixture (and generates a warning).
Fix this at the source by not creating arrays that mix String and
Gstring.