refactor(artifacts): Clean up Spring dependency injection
Each artifact type creates its own okHttpClient; just create a single
one that is autowired into each.

Instead of having the individual types depend on the repository and
modify it, have the repository depend on the individual artifact
types. It can then just create a list from these injected types.

Improve the interface of artifactCredentialsRepository. Most calls
to getAllCredentials filter based on name and/or type; add this
filtering as a supported function in the interface and clean up
the calling code.