fix(elasticsearch): Fix connection string for elasticsearch tests
The elasticsearch tests spin up a new local node agains which to run the
tests; this node defaults to port 9200, but will pick another port if
that port is not available. The tests currently always send requests
to port 9200 (which could be an unrelated node); update them to
send requests to the node created by the tests.