test(core): Simple startup integration test
It's very common for one of our microservices to fail to start
because of a missing autowire dependency. Often this is because
someone changes autowiring such that it only works with a particular
feature enabled, but fails with it disabled.

To prevent these issues, add an extremely simple integration test
that just checks that clouddriver starts up with the default
spring profile. This won't catch everything, but should catch
most of the common cases I've seen. (We can always add test for
other common profiles later if desired.)