refactor(kubernetes): Add optional apiGroup to KubernetesKind
As I was working on the last commit, I found it a bit confusing
knowing whether a given 'kind' string I was dealing with was
a fully-qualified kind (with API group) or just the kind string.

To make this less ambiguous, include 'apiGroup' as a field on
the KubernetesKind object, and handle scoping as part of
toString and fromString so this happens in one place.

This adds the default api group for built-in kinds, but explicitly
considers two built-in kinds as matching even if the API group
doesn't to handle cases such as deployments which can exist in
either apps or extensions.