fix(kubernetes): Infer API version when missing on patch manifest (#3482) (#3483)
* fix(kubernetes): Infer API version when missing on patch manifest

The patch manifest stage is currently failing, because we're fetching
the apiGroup as a required field. If there is no apiGroup specified,
we should try to match the supplied kind against native groups.

* refactor(kubernetes): Add optional apiGroup to KubernetesKind

As I was working on the last commit, I found it a bit confusing
knowing whether a given 'kind' string I was dealing with was
a fully-qualified kind (with API group) or just the kind string.

To make this less ambiguous, include 'apiGroup' as a field on
the KubernetesKind object, and handle scoping as part of
toString and fromString so this happens in one place.

This adds the default api group for built-in kinds, but explicitly
considers two built-in kinds as matching even if the API group
doesn't to handle cases such as deployments which can exist in
either apps or extensions.