feat(core): Add streaming Job functionality
Currently the LocalJobExecutor accumulates all output from the job
into a stream and returns the entire output (as a string) to the
calling function.  In some cases, it would be useful to allow the
calling function to process the output as it is produced to reduce
the amount of output that needs to be stored (and to reduce the
number of times that it is copied around).

Add an optional parameter to runJob specifying a function that will
consume the output as it is produced. When this parameter is passed,
all output will be passed to the supplied callback, and the output
on the JobResult will be null.