refactor(core): Rename JobStatus and member variables
JobStatus is really holding the result of a job; it is never
generated until the job is finished.  Rename it to JobResult.

Now that stdOut can be either stdOut or some processing of the
output, rename it to the more general output.  Also rename
stdErr as error, though either would probably be appropriate for it.