fix(kubernetes): Initialize credentials after CRDs are registered
There are currently two mechanisms for Spinnaker to cache CRDs:
* Registered CRDs that are included in clouddriver's config and
can set CRD-specific properties (such as a handler class)
* Unregistered CRDs that are discovered at startup by asking the
cluster for all CRDs. These use default properties that are not
overriden.

We are currently polling for CRDs in the cluster and configuring
all found CRDs as unregistered before we register configured CRDs.
Registration of a CRD is a no-op if the CRD has already been registered,
so this means the user-configured registered CRD properties are being
ignored.

We should instead wait until we've registered all configured CRDs before
having the caching agent try to auto-discover CRDs.