refactor(provider/kubernetes): Remove reflection from cache constructors
Constructing caching agents currently uses reflection, which breaks
static analysis and makes it harder to reason about and refactor this
code. Instead of having each handler return a reference to the class
that should be constructed, instead have the handler return a functional
interface to the actual constructor so that the base class does not need
to use reflection.