test(kubernetes): Add tests to credentials initializing
There are currently no tests on initializing credentials; before
refactoring this code, add some tests.

Also, a number of the Kubernetes tests depend on being able to explicitly
pass credentials to a builder for KubernetesNamedAccountCredentials,
which will be removed. Instead just use a mock object and specify the
desired properties.