refactor(kubernetes): Remove unused fields from account credentials
Most of the fields on KubernetesNamedAccountCredentials are just
copies of the fields on the contained KubernetesCredentials object.
There's no need to also have these on the parent class (and in
fact we never look these up on the parent class).  Remove all
of the redundant fields.

For the remaining fields:
- Apply 'private final' consistently
- Replace the accessors with a @Getter annotation