refactor(kubernetes): Use a factory class instead of a builder
Now that we're passing the entire ManagedAccount class to the builder
rather than the properties individually, there are no optional
parameters, so a builder class is less optimal. Other than the
ManagedAccount, all properties of the builder are beans that can
be autowired. Create a factory bean that autowires in these other
beans, and exposes methods to create a v1 or a v2 account from a
supplied ManagedAccount.

Ideally most of the logic in CredentialFactory would actually live
in as factories in the credential classes themselves; I'll likely
make this change in a future PR, but for now this is a reasonable
shim to avoid making this PR too large.