fix(kubernetes): Add kube config to test
The V1 provider needs a valid kube config to properly initialize;
ideally we'd be able to inject mocks into the classes but all the
initializing code is using static classes that are called from Java
and can't be mocked. To fix the tests for now, just add a stub
kube config that has the minimum requirements to create the account.