fix(kubernetes): Fail validation if account has no namespaces
Before deploying a manifest, we check whether the namespace is
one that the account is configured for. We do this by checking
namespaces and omitNamespaces, but this fails in the case where
an account is mis-configured and doesn't have any namespaces.

Instead, just use getDeclaredNamespaces which returns a fully resolved
list of namespaces the account knows about (whether because of config
or dynamically discovered).