fix(elasticsearch): Don't autoconfigure elasticsearch/jest (#3644)
We pin our elasticsearch and jest libraries to versions much
older than Spring pulls in. This is currently breaking the
Elasticsearch health check, so disable it by exluding autoconf
of ElasticSearchJestHealthIndicatorAutoConfiguration.

Given that we configure elasticsearch ourselves, we should also
disable ElasticsearchAutoConfiguration and JestAutoConfiguration.
These aren't currently broken, but they are liable to break given
the huge version skew between what we have on the classpath and
what Spring expects.