refactor(kubernetes): Improve typing of pod metrics
The pod metrics code converts the metrics to Maps before
using them, which makes it more difficult to use these
metrics. Simplify this code by using the KubernetesPodMetric
class that already exists.

In order to preserve the serialized version stored in the
cache, rename some fields to match what we're currently
using to serialize. Then serialize/deserialize to/from
that object when we write/read from the cache.