refactor(kubernetes): Consolidate constructors
The multiple constructors in KubernetesKind with various parameters
make it complex to understand which constructor is being called and
what parameters are being defaulted.

Consolidate down to two constructors: one that defaults isDynamic and
isRegistared and is used when we create all the static member instances,
and another that accepts these parameters and is used for dynamic
registration.

If we were creating a lot of these outside of the class it might
make sense to use a builder, but given that this is a private
constructor called primarily from the static members in the
class it seems reasonable to use constructors.