fix(runJob/kubernetes): collect pods and status (#3697)
In order for the Deck to display logs in a reliable manner we need
access to any Pod names created by the Job. When collecting the status
of the Job, collect the names and statuses of any Pods owned by the Job.
When displaying logs, Deck will determine the Pod name use when fetching
console output.