fix(artifacts): resolve version of artifacts
Sometimes Artifacts don’t have the version populated. For example, a default Maven artifact in a pipeline may have a reference with a ‘latest.release’ suffix. This allows us a way to look up the actual version used. Name is often unpopulated along with version.
This also finds specific snapshot versions.
Artifact Url was purposefully left out of the CF server group env vars because we don’t actually have enough information to construct it and put it in artifact.location upstream.