fix(kubernetes): Properly substitute characters in key parts
As kubernetes objects can contain the character ':' which conflicts with
our default cache key separator, the caching logic replaces ':' with ';'
before generating a cache key and then inverts this by replacing ';'
with ':' when splitting a cache key into its component parts.

The current logic that runs when reading cache keys is an expensive
no-op; it substitutes ';' with ':' in the key part but ignores the
results (as the result of replaceAll is ignored).

In addition to fixing the bug, add a performance optimization. replaceAll
always allocates a new string, regardless of whether it actually
needs to perform a substitution; this means that we're allocating
a new string for every cache key part we read from the cache (~6x the
number of keys we read) even though most keys will not have a ':'.
Perform a relatively cheap contains check before doing the replacement.